Refactoring

+++ contracts/interfaces/IInsurerPool.sol
@@ -74,8 +74,8 @@ struct DemandedCoverage {
 
 struct TotalCoverage {
   uint256 totalCoverable; // total demand that can be covered now (already balanced) - this value is not provided per-insured
-  uint64 usableRounds;
-  uint64 openRounds;
+  uint88 usableRounds;
+  uint88 openRounds;
   uint64 batchCount;
 }
